refactor: remove auth service (moved to users service)
