--- src/services/auth.js
@@ -1,46 +1,0 @@
-import client from 'providers/fetchClient'
-import { gql } from '@apollo/client'
-import { formatIsoString } from 'helpers'
-import { format } from 'date-fns'
-
-export const getUser = () => client.get('/v1/me')
-
-export const login = data => client.post('/v1/users/login', data)
-
-//example of rest query using graphql
-export const userQuery = gql`
-  query getUser {
-    me @rest(type: User, path: "/v1/me") {
-      id
-      name
-      email
-      document
-    }
-  }
-`
-
-// example of a graphql query
-export const GET_USER = gql`
-  query GetUser {
-    user {
-      id
-      name
-      email
-      document
-    }
-  }
-`
-
-export const getUsers = async params => {
-  const { results, pageCount } = await client.get('/v1/users', {
-    params: { ...params, created_at: params?.created_at ? format(params.created_at, 'yyyy-MM-dd') : null }
-  })
-
-  return {
-    pageCount,
-    results: results.map(user => ({
-      ...user,
-      formattedCreatedAt: formatIsoString(user.created_at)
-    }))
-  }
-}
